% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_2d.R
\name{area_2d}
\alias{area_2d}
\title{computes the area of a square}
\usage{
area_2d(l, w)
}
\arguments{
\item{l}{a numeric that represents the length}

\item{w}{a numeric that represents the width}
}
\value{
a numeric that is the area of two numbers
}
\description{
computes the area of a square
}
\examples{
l <- 5
w <- 2
area_2d(l,w)
area_2d(0,0)
area_2d(2,0)
}
